use `processSwapTransaction` in `swap/utils`
