+++ src/swap/components/SwapProvider.tsx
@@ -5,26 +5,15 @@ import {
   useMemo,
   useState,
 } from 'react';
-import {
-  useSendTransaction,
-  useConfig,
-  type BaseError,
-  type Config,
-} from 'wagmi';
-import { waitForTransactionReceipt } from 'wagmi/actions';
-import type { SendTransactionMutateAsync } from 'wagmi/query';
+import { useSendTransaction, useConfig, type BaseError } from 'wagmi';
 import { useSwapBalances } from './useSwapBalances';
+import { processSwapTransaction } from '../utils/processSwapTransaction';
 import { buildSwapTransaction } from '../core/buildSwapTransaction';
 import { isSwapError } from '../core/isSwapError';
 import { getSwapQuote } from '../core/getSwapQuote';
 import { USER_REJECTED_ERROR_CODE } from '../constants';
 import { formatTokenAmount } from '../../utils/formatTokenAmount';
-import type {
-  SwapError,
-  SwapErrorState,
-  SwapContextType,
-  BuildSwapTransaction,
-} from '../types';
+import type { SwapError, SwapErrorState, SwapContextType } from '../types';
 import type { Token } from '../../token';
 import type { Address, TransactionReceipt } from 'viem';
 
@@ -255,63 +244,4 @@ export function SwapProvider({
   });
 
   return <SwapContext.Provider value={value}>{children}</SwapContext.Provider>;
-}
-
-export async function processSwapTransaction({
-  swapTransaction,
-  config,
-  setPendingTransaction,
-  setLoading,
-  sendTransactionAsync,
-  onSuccess,
-}: {
-  swapTransaction: BuildSwapTransaction;
-  config: Config;
-  setPendingTransaction: (value: React.SetStateAction<boolean>) => void;
-  setLoading: (value: React.SetStateAction<boolean>) => void;
-  sendTransactionAsync: SendTransactionMutateAsync<Config, unknown>;
-  onSuccess:
-    | ((txReceipt: TransactionReceipt) => void | Promise<void>)
-    | undefined;
-}) {
-  const { transaction, approveTransaction } = swapTransaction;
-
-  // for swaps from ERC-20 tokens,
-  // if there is an approveTransaction present,
-  // request approval for the amount
-  if (approveTransaction?.data) {
-    setPendingTransaction(true);
-    const approveTxHash = await sendTransactionAsync({
-      to: approveTransaction.to,
-      value: approveTransaction.value,
-      data: approveTransaction.data,
-    });
-    await waitForTransactionReceipt(config, {
-      hash: approveTxHash,
-      confirmations: 1,
-    });
-    setPendingTransaction(false);
-  }
-
-  // make the swap
-  setPendingTransaction(true);
-  const txHash = await sendTransactionAsync({
-    to: transaction.to,
-    value: transaction.value,
-    data: transaction.data,
-  });
-  setPendingTransaction(false);
-
-  // wait for swap to land onchain
-  setLoading(true);
-  const transactionObject = await waitForTransactionReceipt(config, {
-    hash: txHash,
-    confirmations: 1,
-  });
-
-  // user callback
-  const callbackResult = onSuccess?.(transactionObject);
-  if (callbackResult instanceof Promise) {
-    await callbackResult;
-  }
 }
